backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.13.3
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==43.0.1
simple-salesforce==1.12.6
openai==1.54.3
langchain==0.3.7
langchain-openai==0.2.8
pgvector==0.3.6
python-multipart==0.0.12
httpx==0.27.2
structlog==24.4.0
pytest==8.3.3
pytest-asyncio==0.24.0
email-validator==2.2.0
